extract types to own zome
