+++ dnas/bridge/zomes/coordinator/transaction_request/Cargo.toml
@@ -9,9 +9,7 @@ name = "transaction_requests"
 
 [dependencies]
 hdk = { workspace = true }
-
 serde = { workspace = true }
-
 bridge_integrity = { workspace = true }
-
+types = { workspace = true }
 transaction_requests_integrity = {path = "../../integrity/transaction_request"}
