+++ dnas/bridge/zomes/integrity/bridge/Cargo.toml
@@ -10,6 +10,5 @@ name = "bridge_integrity"
 [dependencies]
 hdi = { workspace = true }
 hdk = { workspace = true }
-
 serde = { workspace = true }
-#web3 = "0.18.0"
+types = { workspace = true }
