+++ dnas/bridge/zomes/integrity/bridge/src/authority_list.rs
@@ -1,12 +1,6 @@
 use hdi::prelude::*;
 use super::properties::*;
-
-#[hdk_entry_helper]
-#[derive(Clone, PartialEq)]
-pub struct AuthorityList {
-    pub percentage_for_consensus: u32,
-    pub authorities: Vec<(AgentPubKey, String)>,
-}
+use types::*;
 
 pub fn validate_create_authority_list(
     action: EntryCreationAction,
