+++ dnas/bridge/zomes/integrity/transaction_request/src/lib.rs
@@ -1,8 +1,4 @@
-mod entry;
-pub mod signals;
-pub mod types;
 pub mod utils;
 
-pub use entry::*;
 pub use types::*;
 pub use utils::*;
